requirements.txt
PIL==1.1.7
PyODE==1.2.1
PyOpenGL==3.0.2
argparse==1.2.1
distribute==0.6.38
numpy==1.7.1
protobuf==2.5.0
pygame==1.9.1release
pyzmq==13.1.0
wsgiref==0.1.2
